Manually insert a richlistitem into the results box
This works, except the controller doesn't know an item was inserted, so
that keying through the list leads to an off-by-one error.

The controller doesn't expose setters for its list of results, so it
looks like the right answer _will_ turn out to be using a wrapped
version of the UnifiedComplete search service.